Avatar system
Avatar system tested and fully functional. Can load avatars from files/prefabs and customize to a given user. Only implemented for transhumeral case. Created basic object prefabs for the default/test avatar, need to be replaced by proper models. Hand spawning needs to be implemented, but should be a simple copy of others, with the option of hinge (for wrist) or fixed joint. Closes #10