Kinematic energy experiment basic functionality
Moved the functionality of GridExperiment2019 to the new experiment. GridExperiment2019 will be deprecated once the new KE2020 is working. Still has two bugs: After the first iteration, the ball does not register as touched even though it turns green. Data gathering can be cut at the end. Make it so that it triggers task done 1 second after reaching.